04_everything_is_an_api/07_fastapi/my-project/pyproject.toml
[tool.poetry]
name = "my-project"
version = "0.1.0"
description = ""
authors = ["poonyfatima <fatimazahra9736@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.1"
sqlmodel = "^0.0.21"
psycopg = "^3.2.1"
uvicorn = "^0.30.6"
passlib = "^1.7.4"
python-jose = "^3.3.0"
types-python-jose = "^3.3.4.20240106"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
